requirements.txt
beautifulsoup4==4.5.1
click==6.6
colorlover==0.2.1
czml==0.3.3
eventlet==0.19.0
Flask==0.11.1
Flask-Session==0.3.0
Flask-SocketIO==2.8.1
greenlet==0.4.10
itsdangerous==0.24
Jinja2==2.8
lxml==3.6.4
MarkupSafe==0.23
numpy==1.11.2
plotly==1.12.12
pygeoif==0.6
pyproj==1.9.5.1
python-dateutil==2.6.0
python-engineio==1.1.0
python-socketio==1.6.1
pytz==2016.10
redis==2.10.5
requests==2.12.3
six==1.10.0
Werkzeug==0.11.11
